Cleaned up Application Layer

+++ AdLerBackend.Application/Course/CourseManagement/DeleteCourse/DeleteCourseHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AdLerBackend.Application.Common.DTOs.Storage;
+using AdLerBackend.Application.Common.Exceptions;
 using AdLerBackend.Application.Common.Interfaces;
 using AdLerBackend.Application.Common.InternalUseCases.CheckUserPrivileges;
+using AdLerBackend.Application.Moodle.GetUserData;
 using MediatR;
 
 namespace AdLerBackend.Application.Course.CourseManagement.DeleteCourse;
@@ -27,8 +29,21 @@ public class DeleteCourseHandler : IRequestHandler<DeleteCourseCommand, bool>
             WebServiceToken = request.WebServiceToken
         }, cancellationToken);
 
+        var authorData = await _mediator.Send(new GetMoodleUserDataCommand
+        {
+            WebServiceToken = request.WebServiceToken
+        }, cancellationToken);
+
         // get course from db
         var course = await _courseRepository.GetAsync(request.CourseId);
+
+        if (course == null)
+            throw new NotFoundException("Course With Id: " + request.CourseId + " Not Found");
+
+
+        if (course.AuthorId != authorData.UserId)
+            throw new UnauthorizedAccessException("The Course does not belong to the User");
+
 
         // Delete from file System
         _fileAccess.DeleteCourse(new CourseDeleteDto
